requirements.txt
aiogram==3.15.0
aiohttp==3.10.11
APScheduler==3.10.4
python-dotenv==1.0.1
matplotlib==3.9.2
pandas==2.2.3
